Add files via upload
The code below is the code that I have used to plot the Gene Ontology. The data that was used for this code is data which was extracted from the paper “Impaired Muscle Mitochondrial Function in Familial Partial Lipodystrophy (Simha, V. et al 2022)” which identifies the altered gene expression of certain genes present in skeletal muscles. The genes identified to have been altered were inserted into the website pantherdb.org, which allowed me to register them against a database of human genes, ultimately aiming to identify their function within skeletal muscle. The data acquired was then exported and transformed to an excel file, which was then imported into the software R. Looking at the code, we can see that a the x axis reflects the function of the gene. The “gene count” corresponds to the number of genes that are related to this specific function in the reference list used, in our case in Homo Sapiens. The number on the y axis determines the number of genes found to correspond to that function in the dataset that we have. The dots seen are shaded depending on their False Discovery Rate (FDR), which more accurately represents the P Value for our given dataset.